findme/requirements.txt
decorator==4.0.9
Django==1.9.4
django-appconf==1.0.1
django-crispy-forms==1.6.0
django-filer==1.1.1
django-filter==0.13.0
django-floppyforms==1.6.1
django-geojson==2.9.0
django-imagekit==3.3
django-leaflet==0.18.1
django-mptt==0.8.4
django-polymorphic==0.8.1
django-recaptcha==1.0.5
easy-thumbnails==2.3
flake8==2.5.4
geopy==1.11.0
googlemaps==2.4.3
h==0.1.5
Jinja2==2.6
Markdown==2.3.1
mccabe==0.4.0
numpy==1.11.0
oauthlib==1.0.3
path.py==8.1.2
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.6
pilkit==1.1.13
Pillow==3.2.0
psycopg2==2.6.1
ptyprocess==0.5.1
pyflakes==1.0.0
pygdal==1.10.1.0
PyJWT==1.4.0
PyYAML==3.10
requests==2.9.1
requests-oauthlib==0.6.1
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
Unidecode==0.4.19
